adding common code for produce_prom_metric function and implementing it into repack ls
